workitem 284907 Created release build EOCore 2.0.0.147.
April 2018 Release
Tealeaf Module 10.2.1.242, EOCore 2.0.0.147 and TeaCuts 2.0.0.28
Fixed an issue with font color on AbsSpinner.
Fixed an issue with font color on checkbox.
Fixed an issue with RippleDrawable images that use a differnt image format that produced black borders that missing the transparency.
Fixed an issue with image size and placement.
Fixed an issue with get style for drawable of track.
Fixed some issues with color values.
Improved Event ordering.
Fixed an Activity leak issue.